- fixed offset type: the offset was a uint16_t which was causing overwritten data when going past 65536. now it's uint32_t. Bad boy! You're a bad, bad boy!

+++ wav2flash.cpp
@@ -59,7 +59,7 @@ void fillEmptyFile(FILE *_fileStream, uint32_t _fileSize);
 
 
 uint32_t totalWaveFiles;
-uint16_t soundDataOffset;
+uint32_t soundDataOffset;
 int main(int argc, char *argv[])
 {
 	printf("\n\n");
